Fields edited in meals app , nutrition system created

+++ meal_details/models.py
@@ -4,9 +4,9 @@ class MealDetail100g(models.Model) :
     id = models.AutoField(primary_key=True)
     name = models.CharField(max_length=50,unique=True)
     image = models.ImageField(upload_to='meals')
-    protein = models.IntegerField()
-    fats = models.IntegerField()
-    carbs = models.IntegerField()
-    calories = models.IntegerField()
+    protein = models.FloatField()
+    fats = models.FloatField()
+    carbs = models.FloatField()
+    calories = models.FloatField()
     def __str__(self):
         return self.name
